Optifine 5

+++ src/main/java/net/optifine/CustomColorFader.java
@@ -34,7 +34,7 @@ public class CustomColorFader
                 }
                 else
                 {
-                    double d0 = j * 0.001D;
+                    double d0 = (double)j * 0.001D;
                     d0 = Config.limit(d0, 0.0D, 1.0D);
                     double d1 = x - this.color.xCoord;
                     double d2 = y - this.color.yCoord;
